Move new structs to data
